change models

+++ app/src/main/kotlin/model/Main.kt
@@ -111,6 +111,6 @@ class Main {
      */
     @SerializedName("temp_kf")
     @Expose
-    var tempKf: Int? = null
+    var tempKf: Double? = null
 
 }
